+++ spec/factories/transaction.rb
@@ -9,7 +9,14 @@ FactoryBot.define do
     value { 110.01 }
     disbursement_channel { "1" }
     currency { "gbp" }
-    association :provider, factory: :organisation
-    association :receiver, factory: :organisation
+
+    # Government organisation
+    providing_organisation_name { "Department for Business, Energy & Industrial Strategy" }
+    providing_organisation_reference { "GB-GOV-13" }
+    providing_organisation_type { "10" }
+    # Private organisation
+    receiving_organisation_name { Faker::Company.name }
+    receiving_organisation_reference { "GB-COH-{#{Faker::Number.number(digits: 6)}}" }
+    receiving_organisation_type { "70" }
   end
 end